# Fix long execution times caused by isAdRelated