--- a/lighthouse-plugin-publisher-ads/utils/resource-classification.js
+++ b/lighthouse-plugin-publisher-ads/utils/resource-classification.js
@@ -402,6 +402,11 @@ function getNameOrTld(url) {
  */
 function isAdRelated(requestOrUrl) {
   const url = typeof requestOrUrl == 'string' ? requestOrUrl : requestOrUrl.url;
+
+  if (!url) {
+    return false;
+  }
+
   if (isAdScript(url) || getHeaderBidder(url)) {
     return true;
   }