@tollmanz 2022-02-17
I've executed the smoke tests locally with and without the fix and I get the same failures. Seems that something else may have previously introduced a regression in these tests. I've not had time to take a closer look, but should be able to later today.