@tollmanz 2022-02-17
After spending some time with this, I seeing the following:

* `cumulative-ad-shift-0` is failing its initial attempts on master and this branch regularly, but eventually, the 2nd or 3rd attempt does pass. It _feels_ like a race condition. I'm noticing that the first few tries at this test generally fail. The Chrome browser launches, loads the page and quickly exits. When the tests pass, Chrome stays open for a few seconds longer. I'm not sure if there is some condition that the test is waiting for before collecting data and closing the browser, but it seems like it, especially because the errors tend to indicate that there are fewer shifts than expected even when I can visually see the expected errors
* `gpt-errors-overall-0` is failing on master and this branch consistently. The output suggests that a console warning should be thrown to indicate that an invalid size was requested. The actual output shows no console messages were thrown. When running the page locally, I see no console messages. I'm thinking that this could be due to a change in the GPT library