@jburger424 2022-02-17
Thanks for the contribution Zack! Really appreciate the time you spent putting together such a thorough analysis and finding such a simple fix :) This is really quite an impressive delta in time!

It does appear that the failures are not due to your test, I'll update the test assertions later today to unblock this.

We definitely want to credit you in the release notes for this valuable fix, please confirm that that would be ok.