@tollmanz 2022-02-17
> It does appear that the failures are not due to your test, I'll update the test assertions later today to unblock this.

Excellent! We are excited to get this running in our test suite.

> We definitely want to credit you in the release notes for this valuable fix, please confirm that that would be ok.

Definitely ok! Thanks!