@jburger424 2022-02-18
Updated the tests in https://github.com/googleads/publisher-ads-lighthouse-plugin/pull/333, I went ahead and merged the changes in here.