@josephrussell-cn 2022-03-15
@jburger424 thanks for merging this PR. We are looking to integrate the publisher ads plugin into our lighthouse workflow and would like for it to include these performance improvements as it is a non-trivial impact on the execution time without it. When do you anticipate a new version will be published to npm that includes this change? Thank you!